chore: organize graphql querues
